If you are looking at this commit, please read:
This is an empty commit, merely for documentation.
If you are looking at this message then you are
probably curious about my code, that is awesome!

But before you delve deeper please take into consideration
that this code requires some clean-up of unused code.
I am aware of this and of the somewhat poor structure of
the controllers. However, the project works fines like this
and manages all of its resources properly, like database
relationships, making sure no data is leaked, and the like.

For now, the code will remain a little ugly and
disorganized, so please do not be rude to the code,
you might scare the code and it might stop working.

Thanks!